pyproject.toml
[tool.poetry]
name = "poetry_test"
version = "0.1.0"
description = ""
authors = ["renmurakami <autoeyy@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "^9.1.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
